// @flywave/flywave-topo/src/bspline/bezier-curve-base.d.ts
import { CurvePrimitive } from "../curve/curve-primitive";
import { type LineString3d } from "../curve/line-string3d";
import { OffsetOptions } from "../curve/offset-options";
import { StrokeOptions } from "../curve/stroke-options";
import { type IStrokeHandler } from "../geometry3d/geometry-handler";
import { type Plane3dByOriginAndUnitNormal } from "../geometry3d/plane3d-by-origin-and-unit-normal";
import { type Vector3d, Point3d } from "../geometry3d/point3d-vector3d";
import { type Range1d, type Range3d } from "../geometry3d/range";
import { type Ray3d } from "../geometry3d/ray3d";
import { type Transform } from "../geometry3d/transform";
import { type Point4d } from "../geometry4d/point4d";
import { UnivariateBezier } from "../numerics/bezier-polynomials";
import { Bezier1dNd } from "./bezier1d-nd";
import { type KnotVector } from "./knot-vector";
export declare abstract class BezierCurveBase extends CurvePrimitive {
    readonly curvePrimitiveType = "bezierCurve";
    protected _polygon: Bezier1dNd;
    protected _workData0: Float64Array;
    protected _workData1: Float64Array;
    protected _workPoint0: Point3d;
    protected _workPoint1: Point3d;
    protected constructor(blockSize: number, data: Float64Array);
    reverseInPlace(): void;
    saturateInPlace(knotVector: KnotVector, spanIndex: number): boolean;
    get degree(): number;
    get order(): number;
    get numPoles(): number;
    abstract getPolePoint3d(i: number, point?: Point3d): Point3d | undefined;
    abstract getPolePoint4d(i: number, point?: Point4d): Point4d | undefined;
    setInterval(a: number, b: number): void;
    fractionToParentFraction(fraction: number): number;
    emitStrokes(dest: LineString3d, options?: StrokeOptions): void;
    emitStrokableParts(handler: IStrokeHandler, _options?: StrokeOptions): void;
    copyPolesAsJsonArray(): any[];
    isInPlane(plane: Plane3dByOriginAndUnitNormal): boolean;
    polygonLength(): number;
    startPoint(): Point3d;
    endPoint(): Point3d;
    quickLength(): number;
    abstract extendRange(rangeToExtend: Range3d, transform?: Transform): void;
    protected _workBezier?: UnivariateBezier;
    protected _workCoffsA?: Float64Array;
    protected _workCoffsB?: Float64Array;
    protected allocateAndZeroBezierWorkData(primaryBezierOrder: number, orderA: number, orderB: number): void;
    computeStrokeCountForOptions(options?: StrokeOptions): number;
    abstract clone(): BezierCurveBase;
    cloneTransformed(transform: Transform): BezierCurveBase;
    constructOffsetXY(offsetDistanceOrOptions: number | OffsetOptions): CurvePrimitive | CurvePrimitive[] | undefined;
    clonePartialCurve(fractionA: number, fractionB: number): BezierCurveBase;
    projectedParameterRange(ray: Vector3d | Ray3d, lowHigh?: Range1d): Range1d | undefined;
}