// @flywave/flywave-topo/src/bspline/bspline-curve.d.ts
export declare abstract class BSplineCurve3dBase extends CurvePrimitive {
    readonly curvePrimitiveType = "bsplineCurve";
    protected _bcurve: BSpline1dNd;
    private _definitionData?;
    set definitionData(data: any);
    get definitionData(): any;
    protected constructor(poleDimension: number, numPoles: number, order: number, knots: KnotVector);
    get degree(): number;
    get order(): number;
    get numSpan(): number;
    get numPoles(): number;
    copyKnots(includeExtraEndKnot: boolean): number[];
    setWrappable(value: BSplineWrapMode): void;
    abstract evaluatePointInSpan(spanIndex: number, spanFraction: number, result?: Point3d): Point3d;
    abstract evaluatePointAndDerivativeInSpan(spanIndex: number, spanFraction: number, result?: Ray3d): Ray3d;
    abstract knotToPoint(knot: number, result?: Point3d): Point3d;
    abstract knotToPointAndDerivative(knot: number, result?: Ray3d): Ray3d;
    abstract knotToPointAnd2Derivatives(knot: number, result?: Plane3dByOriginAndVectors): Plane3dByOriginAndVectors;
    fractionToPoint(fraction: number, result?: Point3d): Point3d;
    fractionToPointAndDerivative(fraction: number, result?: Ray3d): Ray3d;
    fractionToPointAnd2Derivatives(fraction: number, result?: Plane3dByOriginAndVectors): Plane3dByOriginAndVectors;
    startPoint(): Point3d;
    endPoint(): Point3d;
    reverseInPlace(): void;
    collectBezierSpans(prefer3dH: boolean): BezierCurveBase[];
    abstract getSaturatedBezierSpan3dOr3dH(spanIndex: number, prefer3dH: boolean, result?: BezierCurveBase): BezierCurveBase | undefined;
    abstract getPolePoint4d(poleIndex: number, result?: Point4d): Point4d | undefined;
    abstract getPolePoint3d(poleIndex: number, result?: Point3d): Point3d | undefined;
    poleIndexToDataIndex(poleIndex: number): number | undefined;
    closestPoint(spacePoint: Point3d, _extend: boolean): CurveLocationDetail | undefined;
    abstract clone(): BSplineCurve3dBase;
    cloneTransformed(transform: Transform): BSplineCurve3dBase;
    clonePartialCurve(fractionA: number, fractionB: number): BSplineCurve3dBase;
    appendPlaneIntersectionPoints(plane: PlaneAltitudeEvaluator, result: CurveLocationDetail[]): number;
    constructOffsetXY(offsetDistanceOrOptions: number | OffsetOptions): CurvePrimitive | CurvePrimitive[] | undefined;
    projectedParameterRange(ray: Vector3d | Ray3d, lowHigh?: Range1d): Range1d | undefined;
}